requirements.txt

aiohttp
python-telegram-bot==13.15
requests
